refactor(ui): migrate auth forms to auth directory
